requirements.txt
pydub
python-dotenv
acrcloud
requests
scdl
tqdm

